refactor: split LinkExpiredPage into LinkExpiredForm and useRegistrationEmailResending
