+++ src/app/auth/registration-email-resending/useRegistrationEmailResending.ts
@@ -1,0 +1,92 @@
+'use client'
+
+import { ChangeEvent, useState } from 'react'
+import { SubmitHandler, useForm } from 'react-hook-form'
+
+import { useRegistrationEmailResendingMutation } from '@/src/store/services/authApi'
+import { CustomerError } from '@/src/store/services/types'
+import { zodResolver } from '@hookform/resolvers/zod'
+import { z } from 'zod'
+
+export const useRegistrationEmailResending = () => {
+  const schema = z.object({
+    email: z
+      .string()
+      .min(1, 'Email is required')
+      .email('Invalid email address')
+      .regex(
+        /^[\w-]+(\.[\w-]+)*@([\w-]+\.)+[a-zA-Z]{2,7}$/,
+        'The email must match the format example@example.com'
+      ),
+  })
+
+  type FormType = z.infer<typeof schema>
+
+  const {
+    clearErrors,
+    formState: { errors },
+    getValues,
+    handleSubmit,
+    register,
+    reset,
+    setError,
+    setValue,
+    trigger,
+    watch,
+  } = useForm<FormType>({
+    defaultValues: {
+      email: '',
+    },
+    resolver: zodResolver(schema),
+  })
+
+  const onChangeHandler = (e: ChangeEvent<HTMLInputElement>) => {
+    const value = e.target.value
+
+    setValue('email', value)
+    if (value) {
+      clearErrors('email')
+    }
+  }
+  const [registrationEmailResending, { isLoading }] = useRegistrationEmailResendingMutation()
+  const [showSuccessMessage, setShowSuccessMessage] = useState(false)
+
+  const onSubmit: SubmitHandler<FormType> = async formData => {
+    try {
+      const registrationEmailResendingData = {
+        baseUrl: 'http://localhost:3000',
+        email: formData.email,
+      }
+
+      await registrationEmailResending(registrationEmailResendingData).unwrap()
+
+      setShowSuccessMessage(true)
+    } catch (err) {
+      const error = err as CustomerError
+      const errorMessage = error.data?.messages[0]
+
+      if (errorMessage?.field === 'email') {
+        setError('email', { message: errorMessage.message, type: 'manual' })
+      }
+    }
+  }
+
+  const handleCloseMessage = () => {
+    setShowSuccessMessage(false)
+    reset()
+  }
+
+  return {
+    errors,
+    getValues,
+    handleCloseMessage,
+    handleSubmit,
+    isLoading,
+    onChangeHandler,
+    onSubmit,
+    register,
+    showSuccessMessage,
+    trigger,
+    watch,
+  }
+}
